Debugging of P module
* change the name of waterqaulity1.py > waterquality_vars.py
* change input of soil bulk density to a 3 layer netcdf, and the variable name in the setting files to bulkdensity (instead of rho1, rho2, rho3).
* Fixing bugs in the phosphorus.py
