+++ cwatm/cwatm_initial.py
@@ -155,7 +155,9 @@ class CWATModel_ini(DynamicModel):
 
         self.runoff_concentration_module.initial()
         self.lakes_res_small_module.initial()
-
+        
+        self.waterquality_module.initial()
+        
         self.routing_kinematic_module.initial()
         if checkOption('includeWaterBodies'):
             self.lakes_reservoirs_module.initWaterbodies()
@@ -168,6 +170,5 @@ class CWATModel_ini(DynamicModel):
 
         self.output_module.initial()
         self.environflow_module.initial()
-        #self.waterquality1.initial()
-        self.waterquality_module.initial()
+
